pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movie-barcode"
version = "0.1.0"
description = "Generate movie barcodes from video files using shot-aware K-means color extraction"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["movie", "barcode", "color", "film", "visualization"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "av>=12.0",
    "numpy>=1.26",
    "pillow>=10.0",
    "scikit-learn>=1.4",
    "typer[rich]>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
moviebarcode = "moviebarcode.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["moviebarcode*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
